requirements.txt
langchain==0.0.58
Flask==2.1.0
openai
faiss-cpu
streamlit
streamlit-chat
